chore: organize the file directory
