+++ src/components/domain/group-maker/home/Absentee.vue
