+++ src/components/domain/news-post-checker/results/ResultTable.vue
