+++ src/components/domain/news-post-checker/students/CreateModal.vue
