+++ src/components/domain/news-post-checker/students/StudentTable.vue
